ints are used by the modules and callbacks
2001-12-20 Alexander Rawass <alexannika@users.sourceforge.net>

	* callbacks use now ints instead of floats
	* modules/*.c use now also ints

+++ modules/launch.c
@@ -2,26 +2,26 @@ module launch {
 
   import random;
 
-  void launch_wave_around_area(object fgname,object faction,object type,object ai,float nr_ships,float r1,float r2,object pos){
+  void launch_wave_around_area(object fgname,object faction,object type,object ai,int nr_ships,float r1,float r2,object pos){
     
 
-    float x=_olist.at(pos,0.0)+(random.random(r1,r2)*random.randomsign());
-    float y=_olist.at(pos,1.0)+(random.random(r1,r2)*random.randomsign());
-    float z=_olist.at(pos,2.0)+(random.random(r1,r2)*random.randomsign());
+    float x=_olist.at(pos,0)+(random.random(r1,r2)*random.randomsign());
+    float y=_olist.at(pos,1)+(random.random(r1,r2)*random.randomsign());
+    float z=_olist.at(pos,2)+(random.random(r1,r2)*random.randomsign());
 
-     _io.printf("launching %f ships type=%s faction=%s around area [%f,%f,%f]\n",nr_ships,type,faction,x,y,z);
+     _io.printf("launching %d ships type=%s faction=%s around area [%f,%f,%f]\n",nr_ships,type,faction,x,y,z);
 
     _unit.launch(fgname,faction,type,ai,nr_ships,x,y,z);
   };
 
 
-  void launch_wave_in_area(object fgname,object faction,object type,object ai,float nr_ships,float radius,object pos){
+  void launch_wave_in_area(object fgname,object faction,object type,object ai,int nr_ships,float radius,object pos){
 
-    float x=_olist.at(pos,0.0)+random.random((0.0-radius)/2.0,radius/2.0);
-    float y=_olist.at(pos,1.0)+random.random((0.0-radius)/2.0,radius/2.0);
-    float z=_olist.at(pos,2.0)+random.random((0.0-radius)/2.0,radius/2.0);
+    float x=_olist.at(pos,0)+random.random((0.0-radius)/2.0,radius/2.0);
+    float y=_olist.at(pos,1)+random.random((0.0-radius)/2.0,radius/2.0);
+    float z=_olist.at(pos,2)+random.random((0.0-radius)/2.0,radius/2.0);
 
-    _io.printf("launching %f ships type=%s faction=%s in area [%f,%f,%f]\n",nr_ships,type,faction,x,y,z);
+    _io.printf("launching %d ships type=%s faction=%s in area [%f,%f,%f]\n",nr_ships,type,faction,x,y,z);
 
     _io.message("game","all","launching new wave of fighters:");
     _io.message("game","all",type);
